requirements.txt
psycopg2-binary==2.9.9
loguru==0.7.2
bcrypt==4.1.3
pandas==2.2.2
pytest==8.2.2
openpyxl==3.1.3
fastapi==0.111.0
python-jose==3.3.0
passlib==1.7.4
asyncpg==0.29.0
celery==5.4.0
redis==5.0.5
pydantic-settings==2.3.1
flower==2.0.1
